Clean up initialization sequence
